Create version specific folders for macros
- update scalafmt to use the scala 2.13 dialect unless it's in a scala-3 directory or is an sbt file
- create version specific directories for Macros and the objects that use them, with the traits/functions that need macros defined within the *UsingMacrosMixin traits. The only exception is needing to define AssertMacros outside of macro-lib since it needs access to the Assert class at compilation time and the macro class doesn't depend on the lib module.
- fix issue with abstract lazy val in NodeInfo, by removing it since it's not needed
- fix test path scala version for scala 3. Since Scala 3 uses the 2.13.16 library, scala.util.Properties.versionNumberString returns the library version, which in Scala 3 is 2.13.16. So we use universal/stage/lib to get the scala library version
- set type as needed
- verify removal of protected defs needed in scala 3
- fix formatting issue with frames map
- replace equals override with matches def since scla 3 doesn't allow comparing a string and DebugCommand using ==

DAFFODIL-2975

+++ daffodil-lib/src/test/scala/org/apache/daffodil/lib/util/TestMaybeTakPerf.scala
@@ -159,7 +159,8 @@ class TestMaybeTakPerf {
 
     println("A takeon is " + TakTimer.takeon + " nanoseconds on this platform.")
 
-    slowdown = takSlowdown(taki(MaybeInt(x), MaybeInt(y), MaybeInt(z)), "MaybeInt")
+    slowdown =
+      takSlowdown(taki(MaybeInt(x.toInt), MaybeInt(y.toInt), MaybeInt(z.toInt)), "MaybeInt")
     println("MaybeInt objects are %s times slower".format(slowdown))
 
     slowdown = takSlowdown(takml(MaybeULong(x), MaybeULong(y), MaybeULong(z)), "MaybeULong")
